+++ NEWS.md
@@ -1,3 +1,10 @@
+# divraster 1.0.4
+
+* New initial test added to ensure species names match.
+[#3](https://github.com/flaviomoc/divraster/issues/3).
+* Minor corrections in function tests to align with BAT update.
+* Citation updated.
+
 # divraster 1.0.3
 
 * Improvements in function descriptions.